+++ CHANGELOG.rst
@@ -2,6 +2,8 @@
 Changelog
 =========
 
+* Added ``hide`` parameter to ElasticsearchTarget
+* Added ExpressionTarget support for ElasticSearch data sources
 * Added ``legendCalcs`` parameter to TimeSeries Panel
 
 0.6.2 (2022-02-24)