+++ CHANGELOG.rst
@@ -2,7 +2,13 @@
 Changelog
 =========
 
+
+0.6.3 (2022-03-30)
+==================
+
+* Added Azure Monitor Target
 * Added ``legendCalcs`` parameter to TimeSeries Panel
+
 
 0.6.2 (2022-02-24)
 ==================